requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
blis==0.2.4
certifi==2023.7.22
charset-normalizer==3.3.1
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==8.1.7
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#sha256=4db16860a8cdef56d436038ace6abeb9181a5176bdc8c16c755a20dced51e5f1
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
joblib==1.3.2
Mako==1.2.4
MarkupSafe==2.1.3
mathparse==0.1.2
murmurhash==1.0.10
nltk==3.8.1
numpy==1.24.4
Pint==0.21.1
plac==0.9.6
preshed==2.0.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==3.13.0
python-dateutil==2.7.5
python-decouple==3.8
pytz==2023.3.post1
PyYAML==5.1.2
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spacy==2.1.9
SQLAlchemy==1.3.24
srsly==1.0.7
starlette==0.27.0
thinc==7.0.8
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
wasabi==0.10.1
zipp==3.17.0
